rename hero image
