+++ frontend/src/components/Auth/PasswordReset.tsx
@@ -7,7 +7,7 @@ import { useNavigate, useSearchParams } from "react-router-dom";
 import { LoaderCircle } from "lucide-react";
 import { useLanguage } from "@/context/LanguageContext";
 import { t } from "@/translations";
-import hero from "@/assets/23045509900_f5dfa32a54_k.jpg";
+import hero from "@/assets/hero.jpg";
 
 const PasswordReset = () => {
   const navigate = useNavigate();
