+++ frontend/src/pages/LandingPage.tsx
@@ -1,4 +1,4 @@
-import hero from "@/assets/23045509900_f5dfa32a54_k.jpg";
+import hero from "@/assets/hero.jpg";
 import { Button } from "@/components/ui/button";
 import { useLanguage } from "@/context/LanguageContext";
 import { t } from "@/translations";
